# api-spec-v3.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/cards':
    post:
      tags:
        - 'UCM2: Card creation'
        - 'Cards / community marketplace'
      summary: "Create a new card to go on the community marketplace"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cardId:
                    type: integer
                    description: The ID of the card that has just been created
                    example: 500
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Returned if a user who is not a GAA tries to create a card for somebody else.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
    get:
      tags:
        - 'UCM2: Card creation'
        - 'Cards / community marketplace'
      summary: 'Retrieve all cards for a given section'
      parameters:
        - name: section
          in: query
          description: The section for which cards should be retrieved.
          required: true
          schema:
            type: string
            enum:
              - ForSale
              - Wanted
              - Exchange

        - name: pageNum
          in: query
          description: On frontend, 1 will be passed to start from first page.
          required: true
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: Number of cards per page.
          required: true
          schema:
            type: integer

        - name: sortBy
          in: query
          description: Can sort by the following attributes; CREATED, TITLE, COUNTRY, and KEYWORDS.
            If the parameter is not passed, the default sorting is CREATED.
          required: false
          schema:
            type: string
            enum:
              - CREATED
              - TITLE
              - COUNTRY
              - KEYWORDS


        - name: reverse
          in: query
          description: False by default, but if true, the list will be sorted in descending order.
          required: false
          schema:
            type: boolean
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a (possibly empty) list of cards is returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  '/cards/{id}':
    get:
      tags:
        - 'UCM3: Marketplace section display'
        - 'Cards / community marketplace'
      summary: 'Retrieve a specific card'
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a card is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    delete:
      tags:
        - 'UCM4: Card expiry'
        - 'Cards / community marketplace'
      summary: Delete a community marketplace card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      responses:
        '200':
          description: Card deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to delete a card that they are not the creator of AND the user is not a GAA'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      tags:
        - 'UCM9: Edit card'
        - 'Cards / community marketplace'
      summary: Modify an existing community marketplace card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - section
                - keywords
              properties:
                section:
                  type: string
                  enum:
                    - ForSale
                    - Wanted
                    - Exchange
                title:
                  type: string
                  example: "1982 Lada Samara"
                description:
                  type: string
                  example: "Beige, suitable for a hen house. Fair condition. Some rust. As is, where is. Will swap for budgerigar."
                keywords:
                  type: string
                  example: "Used, Beige"
      responses:
        '200':
          description: Card updated successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to update a card they do not own AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/cards/{id}/extenddisplayperiod':
    put:
      tags:
        - 'UCM4: Card expiry'
        - 'Cards / community marketplace'
      summary: 'Extend the display period of a card that is nearing expiry'
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: End of display period successfully extended. Extension amount should be a further two weeks.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user who is neither the creator of the card, nor a GAA, tries to extend the display period of the card.'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/users/{id}/cards':
    get:
      tags:
        - 'UCM5: Find my cards'
        - 'Cards / community marketplace'
      summary: 'Retrieve all active cards that a user has created'
      parameters:
        - name: id
          in: path
          description: The id of the user for whom to get cards.
          required: true
          schema:
            type: integer
            example: 100
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a (possibly empty) list of cards is returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
    Keyword:
      type: object
      description: "Represents a single 'keyword' within the system - similar to a hashtag."
      properties:
        id:
          type: integer
          example: 600
          readOnly: true
          description: Unique ID, autogenerated by the system
        name:
          type: string
          description: The text value of the keyword
          example: Vehicle
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the keyword was created
    Card:
      type: object
      description: Represents an item in the Community Marketplace
      required:
        - title
        - creatorId
        - section
        - keywordIds
      properties:
        id:
          type: integer
          example: 500
          readOnly: true
          description: Unique ID, autogenerated by the system
        creatorId:
          type: integer
          example: 100
          writeOnly: true
        creator:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/User'
        section:
          type: string
          enum:
            - ForSale
            - Wanted
            - Exchange
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the card was created
        displayPeriodEnd:
          type: string
          format: date-time
          example: '2021-07-29T05:10:00Z'
          readOnly: true
          description: Date and time when the display period of the card is set to expire. This defaults to 2 weeks from creation, and can be extended per UCM4.
        title:
          type: string
          description: Title of the card, should be suitable for one-line displays.
          example: "1982 Lada Samara"
        description:
          type: string
          description: A more detailed description of the card.
          example: "Beige, suitable for a hen house. Fair condition. Some rust. As is, where is. Will swap for budgerigar."
        keywords:
          type: array
          readOnly: true
          description: Keywords describing the card. A card can have one or more keywords associated with it.
          items:
            $ref: '#/components/schemas/Keyword'
        keywordIds:
          type: array
          writeOnly: true
          description: The IDs of keywords that will be associated with a card once it is created. These keywords are created through the POST /keywords endpoints (UCM6).
          items:
            type: integer
            example: [20, 15, 600]
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: "The ID of the 'primary' administrator of the business. This is automatically given to the creator of the business."
        name:
          type: string
          example: 'Lumbridge General Store'
        description:
          type: string
          example: 'A one-stop shop for all your adventuring needs'
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
            - Administrative and Support services
            - Agriculture, Forestry and Fishing
            - Arts and Recreation Services
            - Construction
            - Education and Training
            - Electricity, Gas, Water and Waste Services
            - Financial and Insurance Services
            - Health Care and Social Assistance
            - Information Media and Telecommunication
            - Manufacturing
            - Mining
            - Professional, Scientific and Technical Services
            - Public Administration and Safety
            - Rental, Hiring and Real Estate Services
            - Transport, Postal and Warehousing
            - Wholesale Trade
            - Other Services
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: '3/24'
        streetName:
          type: string
          example: 'Ilam Road'
        suburb:
          type: string
          example: 'Upper Riccarton'
        city:
          type: string
          example: 'Christchurch'
        region:
          type: string
          example: 'Canterbury'
        country:
          type: string
          example: 'New Zealand'
        postcode:
          type: string
          example: '90210'
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
        phoneNumber:
          type: string
          example: '+64 3 555 0129'
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: '1337-H%nt3r2'
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: '(Implemented in U4): This property should only be shown to Global Application Admins'
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: '(Implemented in U6): A list of businesses for which the user is an administrator'
          items:
            $ref: '#/components/schemas/Business'
...